test : 테스트 형태 변경
